Update model tests to reflect the new expected behaviour
We are now expecting users to only add start and end dates for children who are only being claimed for part of the tax year.
The tests have been updated to reflect this.  Tests that only contained children whose start and end dates matched the tax year have been changed to only pass full year children to the tests.

Add a model test to calculate the benefit amount for multiple full and part year children